Usernames are now enforced to be email addresses. Turned login screen into a component with multiple slots to allow overriding of certain sections
